+++ scripting/typeDef.h
@@ -16,7 +16,7 @@ namespace BraneScript
         virtual ~TypeDef() = default;
 
         virtual const char* name() const = 0;
-        virtual uint32_t size() const = 0;
+        virtual uint16_t size() const = 0;
         virtual ValueType type() const = 0;
     };
 }
